Changed WSAIoctl function to return the right I.P.
AOM tries to access the I.P. of an interface with index 2, but wine has only one interface with index 0. So the "fix" is to force the function to return the I.P. of the interface with index 0, whenever AOM asks for the I.P. of interface 2. Unfortunately, this does fix the problem of the 0.0.0.0 I.P. address showing in lobby, but after a couple of minutes the game crashes..